send on work

+++ src/main/java/com/igorgorbachev/SpringBootBK/service/KlientServiceImpl.java
@@ -44,7 +44,11 @@ public class KlientServiceImpl implements KlientService {
     @Transactional
     @Override
     public void deleteKlient(Klient klient) {
-        klientDao.deleteKlient(klient);
+        if (klient.getCar() != null && !klient.getCar().isEmpty()) {
+            throw new IllegalStateException("Клиента нельзя удалить, так как у него есть автомобили.");
+        } else  {
+            klientDao.deleteKlient(klient);
+        }
     }
 
     @Override
